requirements.txt
bcrypt==4.0.1
click==8.1.3
dnspython==2.3.0
Flask==2.2.3
Flask-PyMongo==2.3.0
itsdangerous==2.1.2
Jinja2==3.1.2
jsonify==0.5
MarkupSafe==2.1.2
pymongo==4.3.3
python-bsonjs==0.3.0
python-dotenv==1.0.0
Werkzeug==2.2.3
